OTP/requirement.txt
pyinstaller
pycryptodome
oathtool